requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# servapin/app.py: 1
# servapin/resources/api.py: 2
# servapin/resources/auth.py: 2
# servapin/resources/servapi.py: 6
Flask == 1.1.2

# servapin/app.py: 2
# servapin/database/models.py: 2
Flask_Bcrypt == 0.7.1

# servapin/app.py: 3
# servapin/resources/api.py: 3
# servapin/resources/auth.py: 3
Flask_JWT_Extended == 4.1.0

# servapin/app.py: 6
# servapin/resources/api.py: 4
# servapin/resources/auth.py: 4
# servapin/resources/servapi.py: 8
Flask_RESTful == 0.3.8

# servapin/resources/api.py: 7
# servapin/resources/servapi.py: 5
WooCommerce == 2.1.1

# servapin/database/db.py: 2
flask_mongoengine == 1.0.0

# servapin/database/models.py: 3
# servapin/database/z_samples/model_product.py: 2
mongoengine == 0.23.0

# servapin/database/db.py: 1
# servapin/resources/api.py: 1
pymongo == 3.11.3

# servapin/resources/api.py: 6
# servapin/resources/servapi.py: 2
python_decouple == 3.4

# servapin/app.py: 9
# servapin/resources/servapi.py: 7
waitress == 1.4.4
